+++ spotdl/types/song.py
@@ -113,7 +113,7 @@ class Song:
             genres=raw_album_meta["genres"] + raw_artist_meta["genres"],
             disc_number=raw_track_meta["disc_number"],
             disc_count=int(raw_album_meta["tracks"]["items"][-1]["disc_number"]),
-            duration=raw_track_meta["duration_ms"] / 1000,
+            duration=int(raw_track_meta["duration_ms"] / 1000),
             year=int(raw_album_meta["release_date"][:4]),
             date=raw_album_meta["release_date"],
             track_number=raw_track_meta["track_number"],
